spv_admin/pyproject.toml
[project]
name = "spv"
version = "0.1.0"
description = "SPV implementation for the RWA infrastructure, this system was designed to run in a offline environmant (localhost + internet)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioipfs>=0.7.1",
    "ape-alchemy>=0.8.10",
    "ape-avalanche>=0.8.1",
    "ape-optimism>=0.8.3",
    "ape-vyper>=0.8.9",
    "django>=6.0.1",
    "eth-ape>=0.8.45",
    "python-decouple>=3.8",
    "tenacity>=9.1.2",
]
